# schemas/openapi-commercial-v1.in.yaml
openapi: "3.0.3"
info:
  title: AgentSkeptic commercial license API
  version: "1.0.0"
  contact:
    url: __DISTRIBUTION_CONTACT_URL__
  x-agentskeptic-distribution:
    repository: __DISTRIBUTION_REPO_URL__
    npmPackage: __DISTRIBUTION_NPM_URL__
    openApi: __OPENAPI_SELF_URL__
  description: "__IDENTITY_ONE_LINER__\n\nMachine-readable contract for license preflight used by the published npm CLI.\nBase URL is your deployed app origin (same as NEXT_PUBLIC_APP_URL).\n\nEvery path in this document returns an `x-request-id` response header on all status codes (echo a valid client `x-request-id` when supplied; otherwise server-generated). Non-2xx JSON bodies follow RFC 7807-style Problem Details (`type`, `title`, `status`, `detail`, optional `code`, `instance`) unless noted; `POST /api/v1/usage/reserve` denials additionally include legacy `allowed`, `code`, `message`, and optional `upgrade_url` for backward compatibility.\n"
...
servers:
  - url: __SERVERS_ORIGIN__
    description: Replace with your deployment origin
paths:
...
  /api/v1/funnel/verify-outcome:
    post:
      operationId: postVerifyOutcomeBeacon
      summary: Licensed verify-outcome beacon (idempotent per API key + run_id)
      description: >
        POST after a successful `POST /api/v1/usage/reserve` for the same `run_id`. Success is HTTP 204 with an empty body.
        Requires Bearer API key. Body schemaVersion must be 2.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOutcomeRequestV2"
      responses:
        "204":
          description: Beacon accepted (or duplicate ignored)
        "400":
          description: Invalid JSON or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No reservation for this run_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "410":
          description: Reservation too old
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/oss/claim-ticket:
    post:
      operationId: postOssClaimTicket
      summary: Mint OSS claim ticket (CLI only; product activation headers required)
      description: >
        Creates or returns a handoff URL for linking a verification run to a web account.
        Requires `X-AgentSkeptic-Cli-Product` and `X-AgentSkeptic-Cli-Version` headers (see website/src/lib/funnelProductActivationConstants.ts).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssClaimTicketRequest"
      responses:
        "200":
          description: Handoff URL returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OssClaimTicketHandoffResponse"
        "204":
          description: Ticket already claimed (idempotent)
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (unsupported CLI client)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
    VerifyOutcomeRequestV2:
      type: object
      required:
        - schema_version
        - run_id
        - workflow_id
        - trust_decision
        - reason_codes
        - terminal_status
        - workload_class
        - subcommand
      properties:
        schema_version:
          type: integer
          const: 2
        run_id:
          type: string
          maxLength: 256
        workflow_id:
          type: string
          maxLength: 512
        trust_decision:
          type: string
          enum: [safe, unsafe, unknown]
        reason_codes:
          type: array
          maxItems: 8
          items:
            type: string
        terminal_status:
          type: string
          enum: [complete, inconsistent, incomplete]
        workload_class:
          type: string
          enum: [bundled_examples, non_bundled]
        subcommand:
          type: string
          enum: [batch_verify, quick_verify, verify_integrator_owned]
    OssClaimTicketRequest:
      type: object
      description: v1 or v2 OSS claim ticket wire shape (see website/src/lib/ossClaimTicketPayload.ts).
      additionalProperties: true
    OssClaimTicketHandoffResponse:
      type: object
      required: [schema_version, handoff_url]
      properties:
        schema_version:
          type: integer
          const: 2
        handoff_url:
          type: string
          format: uri
...